waittime: now can keep bigger values, size was too small before - patch provided by CDWriter
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3545 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/sql/mangos.sql
@@ -976,7 +976,7 @@ CREATE TABLE `creature_movement` (
   `position_x` float NOT NULL default '0',
   `position_y` float NOT NULL default '0',
   `position_z` float NOT NULL default '0',
-  `waittime` tinyint(3) unsigned NOT NULL default '0',
+  `waittime` int(5) unsigned NOT NULL default '0',
   `text1` varchar(255) default NULL,
   `text2` varchar(255) default NULL,
   `text3` varchar(255) default NULL,
